Remove auto version grab to fix package deployment
The trick used herein to set the version from git tags is not only run
from within this repo: it is also run on deployment sites, by pipenv
and/or pip. Therefore, let's stick to fixed strings and update it by
hand like peasants until we figure how to build setup.py from a
template.