Fix: no dequeue/rejoin when lacking instances
When requesting to take multiple instances, a process X would be queued
if not enough instances were available. Then, every time another process
would release the instances it had, X would pop out of the queue, but
would be made to join it back if the number of free instances was
insufficient. With a default queue, this would make X go back to the
end, breaking the resource access queueing semantics.

The fix involves adding a "tag" of information to queued processes, and
a peek() method to queues, to check the tag of the queue top. For
resources, the number of requested instances is put up as the tag: this
allows to check for sufficient availability before popping the awaiting
process.

Also: removed a would-be unit test that resulted in deadlock of
concurrent processes.