Fix resource leak due to circular reference system
A simulator refers to events pertaining to all hanging processes, and
processes refer to the simulator they run into -- a system of circular
references. For the processes, the simulator reference is now held
weakly. This allows the simulator to be GC'ed, through which it triggers
explicit teardown of hanging proceses. Thus all ends up reclaimed once a
simulator's "owning" reference is discarded.